Remove bind app and migrate entity base
