--- bind/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class BindConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "bind"
-    verbose_name = "Bind Models"
