--- bind/fixtures/request_types.json
@@ -1,11 +1,0 @@
-[
-  {
-    "model": "bind.requesttype",
-    "pk": 1,
-    "fields": {
-      "code": "GEN",
-      "name": "General",
-      "next_number": 30000
-    }
-  }
-]
