--- bind/meta/__init__.py
@@ -1,1 +1,0 @@
-"""Meta integration utilities."""
