--- bind/templates/admin/seed_data.html
@@ -1,35 +1,0 @@
-{% extends "admin/base_site.html" %}
-{% load i18n %}
-
-{% block content %}
-<h1>Seed data</h1>
-<table class="adminlist">
-  <thead>
-    <tr>
-      <th>Model</th>
-      <th>Object</th>
-      <th>Status</th>
-      <th></th>
-    </tr>
-  </thead>
-  <tbody>
-  {% for item in seed_items %}
-    <tr>
-      <td>{{ item.model_verbose_name }}</td>
-      <td>{{ item.obj }}</td>
-      <td>{% if item.obj.is_deleted %}Deleted{% else %}Active{% endif %}</td>
-      <td>
-        {% if item.obj.is_deleted %}
-        <form method="post" style="display:inline">{% csrf_token %}
-          <input type="hidden" name="app" value="{{ item.model_app_label }}">
-          <input type="hidden" name="model" value="{{ item.model_name }}">
-          <input type="hidden" name="pk" value="{{ item.obj.pk }}">
-          <button type="submit" class="button">Reinstall</button>
-        </form>
-        {% endif %}
-      </td>
-    </tr>
-  {% endfor %}
-  </tbody>
-</table>
-{% endblock %}
