--- bind/urls.py
@@ -1,12 +1,0 @@
-from django.urls import path
-
-from . import views
-
-app_name = "bind"
-
-urlpatterns = [
-    path("register/", views.register, name="register"),
-    path("post/", views.post, name="post"),
-    path("domain-post/", views.domain_post, name="domain-post"),
-    path("odoo/test/<int:pk>/", views.test_connection, name="odoo-test"),
-]
