+++ pages/templates/admin/index.html
@@ -88,9 +88,6 @@
 {% block content_title %}
 <div id="admin-home-header" style="display:flex; justify-content:space-between; align-items:center;">
   <h1>Home</h1>
-  <div>
-    <a href="{% url 'admin:seed-data' %}" class="button">Seed data</a>
-  </div>
  </div>
 {% endblock %}
