+++ pyproject.toml
@@ -9,7 +9,7 @@ description = "Django-based MESH system"
 requires-python = ">=3.10"
 license = "MIT"
 classifiers = [ "Programming Language :: Python :: 3", "Framework :: Django",]
-dependencies = [ "amqp==5.3.1", "annotated-types==0.7.0", "anyio==4.9.0", "asgiref==3.9.1", "atproto==0.0.61", "attrs==25.3.0", "autobahn==24.4.2", "Automat==25.4.16", "billiard==4.2.1", "bleach==6.2.0", "celery==5.5.3", "certifi==2025.7.14", "cffi==1.17.1", "channels==4.1.0", "charset-normalizer==3.4.2", "click==8.2.1", "click-didyoumean==0.3.1", "click-plugins==1.1.1.2", "click-repl==0.3.0", "colorama==0.4.6", "constantly==23.10.4", "cron-descriptor==1.4.5", "cryptography==45.0.5", "daphne==4.2.1", "diff-match-patch==20241021", "Django==5.2.4", "django-celery-beat==2.8.1", "django-fernet-fields==0.6", "django-import-export==4.3.9", "django-post_office==3.10.1", "django-timezone-field==7.1", "dnspython==2.7.0", "docutils==0.22", "gpiozero==2.0.1; sys_platform == \"linux\"", "h11==0.16.0", "httpcore==1.0.9", "httpx==0.28.1", "hyperlink==21.0.0", "idna==3.10", "incremental==24.7.2", "kombu==5.5.4", "libipld==3.1.1", "Markdown==3.8.2", "mfrc522==0.0.7; sys_platform == \"linux\"", "outcome==1.3.0.post0", "packaging==25.0", "pillow==11.3.0", "prompt_toolkit==3.0.51", "psycopg==3.2.9", "psycopg-binary==3.2.9", "pyasn1==0.6.1", "pyasn1_modules==0.4.2", "pycparser==2.22", "pydantic==2.11.7", "pydantic_core==2.33.2", "pyOpenSSL==25.1.0", "pyperclip==1.9.0", "PySocks==1.7.1", "python-crontab==3.3.0", "python-dateutil==2.9.0.post0", "python-dotenv==1.1.1", "qrcode==8.2", "requests==2.32.4", "selenium==4.34.2", "service-identity==24.2.0", "setuptools==80.9.0", "six==1.17.0", "smbus2==0.5.0", "sniffio==1.3.1", "sortedcontainers==2.4.0", "sqlparse==0.5.3", "tablib==3.8.0", "tinycss2==1.4.0", "toml==0.10.2", "trio==0.30.0", "trio-websocket==0.12.2", "Twisted==25.5.0", "txaio==25.6.1", "typing-inspection==0.4.1", "typing_extensions==4.14.1", "tzdata==2025.2", "urllib3==2.5.0", "vine==5.1.0", "wcwidth==0.2.13", "webencodings==0.5.1", "websocket-client==1.8.0", "websockets==13.1", "wsproto==1.2.0", "zope.interface==7.2"]
+    dependencies = [ "amqp==5.3.1", "annotated-types==0.7.0", "anyio==4.9.0", "asgiref==3.9.1", "atproto==0.0.61", "attrs==25.3.0", "autobahn==24.4.2", "Automat==25.4.16", "billiard==4.2.1", "bleach==6.2.0", "celery==5.5.3", "certifi==2025.7.14", "cffi==1.17.1", "channels==4.1.0", "charset-normalizer==3.4.2", "click==8.2.1", "click-didyoumean==0.3.1", "click-plugins==1.1.1.2", "click-repl==0.3.0", "colorama==0.4.6", "constantly==23.10.4", "cron-descriptor==1.4.5", "cryptography==45.0.5", "daphne==4.2.1", "diff-match-patch==20241021", "Django==5.2.4", "django-celery-beat==2.8.1", "django-import-export==4.3.9", "django-post_office==3.10.1", "django-timezone-field==7.1", "dnspython==2.7.0", "docutils==0.22", "gpiozero==2.0.1; sys_platform == \"linux\"", "h11==0.16.0", "httpcore==1.0.9", "httpx==0.28.1", "hyperlink==21.0.0", "idna==3.10", "incremental==24.7.2", "kombu==5.5.4", "libipld==3.1.1", "Markdown==3.8.2", "mfrc522==0.0.7; sys_platform == \"linux\"", "outcome==1.3.0.post0", "packaging==25.0", "pillow==11.3.0", "prompt_toolkit==3.0.51", "psycopg==3.2.9", "psycopg-binary==3.2.9", "pyasn1==0.6.1", "pyasn1_modules==0.4.2", "pycparser==2.22", "pydantic==2.11.7", "pydantic_core==2.33.2", "pyOpenSSL==25.1.0", "pyperclip==1.9.0", "PySocks==1.7.1", "python-crontab==3.3.0", "python-dateutil==2.9.0.post0", "python-dotenv==1.1.1", "qrcode==8.2", "requests==2.32.4", "selenium==4.34.2", "service-identity==24.2.0", "setuptools==80.9.0", "six==1.17.0", "smbus2==0.5.0", "sniffio==1.3.1", "sortedcontainers==2.4.0", "sqlparse==0.5.3", "tablib==3.8.0", "tinycss2==1.4.0", "toml==0.10.2", "trio==0.30.0", "trio-websocket==0.12.2", "Twisted==25.5.0", "txaio==25.6.1", "typing-inspection==0.4.1", "typing_extensions==4.14.1", "tzdata==2025.2", "urllib3==2.5.0", "vine==5.1.0", "wcwidth==0.2.13", "webencodings==0.5.1", "websocket-client==1.8.0", "websockets==13.1", "wsproto==1.2.0", "zope.interface==7.2"]
 [[project.authors]]
 name = "Rafael J. Guillén-Osorio"
 email = "tecnologia@gelectriic.com"
@@ -23,4 +23,4 @@ Repository = "https://github.com/arthexis/arthexis"
 Homepage = "https://arthexis.com"
 
 [tool.setuptools]
-packages = [ "core", "config", "nodes", "ocpp", "pages", "bind",]
+    packages = [ "core", "config", "nodes", "ocpp", "pages",]
