+++ requirements.txt
@@ -25,7 +25,6 @@ daphne==4.2.1
 diff-match-patch==20241021
 Django==5.2.4
 django-celery-beat==2.8.1
-django-fernet-fields==0.6
 django-import-export==4.3.9
 django-post_office==3.10.1
 django-timezone-field==7.1
